+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,19 +1,14 @@
 package roomescape.controller;
 
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import roomescape.domain.Reservation;
+import roomescape.dto.ReservationRequest;
+import roomescape.dto.ReservationResponse;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicLong;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.ReservationRequest;
-import roomescape.dto.ReservationResponse;
-import roomescape.model.Reservation;
 
 @RestController
 @RequestMapping("/reservations")
@@ -27,7 +22,7 @@ public class ReservationController {
         List<ReservationResponse> reservationResponses = reservations.stream()
                 .map(reservation -> ReservationResponse.from(reservation))
                 .toList();
-        
+
         return ResponseEntity.ok(reservationResponses);
     }
